Split agent prompt helpers
